Move Function timeOfDay
